:sparkles: FinishRunningUseCase 구현

+++ data/src/main/java/com/whyranoid/data/running/RunnerDataSourceImpl.kt
@@ -2,6 +2,7 @@ package com.whyranoid.data.running
 
 import com.google.firebase.firestore.FieldValue
 import com.google.firebase.firestore.FirebaseFirestore
+import com.whyranoid.data.constant.CollectionId
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.callbackFlow
@@ -11,7 +12,7 @@ import kotlin.coroutines.suspendCoroutine
 class RunnerDataSourceImpl(private val db: FirebaseFirestore) : RunnerDataSource {
 
     override fun getCurrentRunnerCount(): Flow<Int> = callbackFlow {
-        db.collection("Runners")
+        db.collection(CollectionId.RUNNERS_COLLECTION)
             .document("runnersId")
             .addSnapshotListener { snapshot, _ ->
                 snapshot?.let {
@@ -26,7 +27,7 @@ class RunnerDataSourceImpl(private val db: FirebaseFirestore) : RunnerDataSource
     override suspend fun startRunning(uid: String): Boolean {
         if (uid.isBlank()) return false
         return suspendCoroutine { continuation ->
-            db.collection("Runners")
+            db.collection(CollectionId.RUNNERS_COLLECTION)
                 .document("runnersId")
                 .update(uid, uid)
                 .addOnSuccessListener {
@@ -39,9 +40,9 @@ class RunnerDataSourceImpl(private val db: FirebaseFirestore) : RunnerDataSource
     }
 
     override suspend fun finishRunning(uid: String): Boolean {
-        // TODO 구현
+        if (uid.isBlank()) return false
         return suspendCoroutine { continuation ->
-            db.collection("Runners")
+            db.collection(CollectionId.RUNNERS_COLLECTION)
                 .document("runnersId")
                 .update(uid, FieldValue.delete())
                 .addOnSuccessListener {
